# openapi/tktsync.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TktSync API",
    "version": "1.0.0",
    "description": "Authoritative TktSync API contract for implemented M3, M4, and M5 Admin and Partner surfaces."
  },
  "paths": {
...
    "/api/v1/partner/events/{event_id}": {
      "get": {
        "operationId": "partnerGetEvent",
        "security": [
          {
            "PartnerBearer": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerEvent"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "description": "Optional client request/correlation UUID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/partner/events/{event_id}/layout": {
      "get": {
        "operationId": "partnerGetEventLayout",
        "security": [
          {
            "PartnerBearer": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObject"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "description": "Optional client request/correlation UUID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/partner/events/{event_id}/availability": {
      "get": {
        "operationId": "partnerGetAvailability",
        "security": [
          {
            "PartnerBearer": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerAvailability"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "description": "Optional client request/correlation UUID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/partner/reservations": {
      "post": {
        "operationId": "partnerCreateReservation",
        "security": [
          {
            "PartnerBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Durable idempotency key scoped to the authenticated Partner and operation.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "description": "Optional client request/correlation UUID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reservation created atomically.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateReservationResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/partner/reservations/{reservation_id}": {
      "get": {
        "operationId": "partnerGetReservation",
        "security": [
          {
            "PartnerBearer": []
          }
        ],
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^res_"
            }
          },
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "description": "Optional client request/correlation UUID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authoritative Reservation state for the owning Partner.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "partnerModifyReservation",
        "security": [
          {
            "PartnerBearer": []
          }
        ],
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^res_"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Durable idempotency key scoped to the authenticated Partner and operation.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-TktSync-Reservation-Token",
            "in": "header",
            "required": true,
            "description": "Opaque Reservation continuation capability. Never place this value in a URL.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "description": "Optional client request/correlation UUID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyReservationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reservation modified atomically.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/partner/reservations/{reservation_id}/checkout": {
      "post": {
        "operationId": "partnerBeginReservationCheckout",
        "security": [
          {
            "PartnerBearer": []
          }
        ],
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^res_"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Durable idempotency key scoped to the authenticated Partner and operation.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-TktSync-Reservation-Token",
            "in": "header",
            "required": true,
            "description": "Opaque Reservation continuation capability. Never place this value in a URL.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "description": "Optional client request/correlation UUID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Protected checkout window established before Partner payment.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeginCheckoutResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/partner/reservations/{reservation_id}/payment-failure": {
      "post": {
        "operationId": "partnerReportReservationPaymentFailure",
        "security": [
          {
            "PartnerBearer": []
          }
        ],
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^res_"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Durable idempotency key scoped to the authenticated Partner and operation.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-TktSync-Reservation-Token",
            "in": "header",
            "required": true,
            "description": "Opaque Reservation continuation capability. Never place this value in a URL.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "description": "Optional client request/correlation UUID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentFailureRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Definitive non-charge recorded and Reservation moved to retry or released.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentFailureResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/partner/reservations/{reservation_id}/release": {
      "post": {
        "operationId": "partnerReleaseReservation",
        "security": [
          {
            "PartnerBearer": []
          }
        ],
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^res_"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Durable idempotency key scoped to the authenticated Partner and operation.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-TktSync-Reservation-Token",
            "in": "header",
            "required": true,
            "description": "Opaque Reservation continuation capability. Never place this value in a URL.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "description": "Optional client request/correlation UUID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment-safe Reservation released with source-aware inventory restoration.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseReservationResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "PartnerBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "ScannerBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "SelectionBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque, time-bounded BuyerSelectionSession capability."
      }
    },
...
    "schemas": {
...
      "PartnerEvent": {
        "type": "object",
        "required": ["id", "name", "state", "server_time"],
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "state": {
            "type": "string",
            "enum": ["DRAFT", "ON_SALE", "PAUSED", "SALES_CLOSED", "COMPLETED", "CANCELLED"]
          },
          "starts_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "ends_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "sales_open_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "sales_close_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "admission_open_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "admission_close_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "timezone_name": {
            "type": ["string", "null"]
          },
          "server_time": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "headers": {
      "XRequestID": {
        "description": "Request/correlation identifier generated or normalized by TktSync.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    }
  }
}